backend_python/requirements.txt
openai
python-dotenv
fastapi[standard]
python-multipart
uvicorn[standard]
PyPDF2
